Canvas section security + tests

+++ app/controllers/canvas_sections_controller.rb
@@ -3,21 +3,23 @@ if Rails.application.config.feature_map.enabled?(:ideation)
     before_filter :authenticate_user!
     
     def create
-      # TODO: verify current user has edit permissions
-      
       spark = Spark.find(params[:spark_id])
-      # Place under the rest
-      lowest_canvas = spark.canvas_sections.order("y + height DESC").first
-      
-      if lowest_canvas.nil?
-        max_y = 0
+      if current_user.nil? || !current_user.can_edit?(spark)
+        message = 'You don\'t have permission to ideate this spark.' 
       else
-        max_y = lowest_canvas.y + lowest_canvas.height
+        # Place under the rest of the canvases
+        lowest_canvas = spark.canvas_sections.order("y + height DESC").first
+        
+        if lowest_canvas.nil?
+          max_y = 0
+        else
+          max_y = lowest_canvas.y + lowest_canvas.height
+        end
+        
+        section = CanvasSection.new({:spark_id => spark.id, :name => 'New Section', :x => 0, :y => max_y, :width => 300, :height => 150})
       end
       
-      section = CanvasSection.new({:spark_id => spark.id, :name => 'New Section', :x => 0, :y => max_y, :width => 300, :height => 150})
-      
-      if section.valid?
+      if !section.nil? && section.valid?
         section.save
         result = section
       else
@@ -26,26 +28,26 @@ if Rails.application.config.feature_map.enabled?(:ideation)
       end
       
       respond_to do |format|
-        format.html { puts "HTML"; redirect_to spark }
-        format.json { puts "JSON"; render json: result }
+        format.html { redirect_to section.nil? ? root_url : section.spark }
+        format.json { render json: result }
         format.js
       end
     end
     
     def update
-      # TODO: verify current user has edit permissions
-      
       sectionId = params[:sectionId] # eg. "section12"
       sectionId = sectionIdToId(sectionId)
+      message = 'Please specify the section' if sectionId.nil?
+      section = CanvasSection.find(sectionId)
+      message = 'You don\'t have permission to ideate this spark.' if current_user.nil? || !current_user.can_edit?(section.spark)
       
       x = params[:x]
       y = params[:y]
-      message = 'Please specify the section' if sectionId.nil?
+      
       message = 'Please specify the x position' if x.nil?
       message = 'Please specify the y position' if y.nil?
       
-      if message.nil?
-        section = CanvasSection.find(sectionId)
+      if message.nil? # valid input
         section.x = x
         section.y = y
         
@@ -60,24 +62,27 @@ if Rails.application.config.feature_map.enabled?(:ideation)
       result = { :error => message } if !message.nil?
       
       respond_to do |format|
-        format.html { puts "HTML"; redirect_to section.spark }
-        format.json { puts "JSON"; render json: result }
+        format.html { redirect_to section.nil? ? root_url : section.spark }
+        format.json { render json: result }
         format.js
       end
     end
     
     def destroy
-      puts "@@@ INSIDE, params are #{params}"
-      # TODO: verify current user has edit permissions
       sectionId = params[:sectionId] # eg. "section12"
       sectionId = sectionIdToId(sectionId)
       message = 'Please specify the section' if sectionId.nil?
       section = CanvasSection.find(sectionId)
-      section.destroy
+      
+      if current_user.nil? || !current_user.can_edit?(section.spark)
+        message = 'You don\'t have permission to ideate this spark.'
+      else
+        section.destroy
+      end
       
       respond_to do |format|
-        format.html { puts "HTML"; redirect_to section.spark }
-        format.json { puts "JSON"; render json: { success: true, message: message } }
+        format.html { redirect_to section.nil? ? root_url : section.spark }
+        format.json { render json: { success: true, message: message } }
         format.js
       end
     end
